requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-pinecone
pinecone-client
pypdf
beautifulsoup4
huggingface_hub
sentence-transformers
